fix: address /code-review findings on #531
- Fix plugins/iac-skill/skills/iac-authoring/SKILL.md: its egress allowlist
  used bare host strings, which the frontmatter parser silently drops (only
  mapping-node entries survive) - this repo's one shipped manifest that would
  have exercised the new validator, and it was invisible to it. Corrected to
  the structured host/schemes/ports shape the schema already requires.
- Split SkillMetadataParser.cs into a second partial (SkillMetadataParser.Egress.cs)
  to bring the main file back under the 400-line convention.
- Document two known, deliberately out-of-scope gaps in per-skill egress
  scoping - the AIContext.Tools channel (GoverningToolContextProvider) and
  same-named-tool pinning across merged skills (ToolChainBuilder.Surface.cs) -
  tracked as #589.
- Extract the ~23 inline `new EgressManifestValidator()` test call sites into
  a shared TestMcpSecurityScanner.RealEgressValidator() helper, matching the
  existing AlwaysSafe()/DefaultConfig() convention.
- Fix SkillEgressConsumerResolvedExemptions_AreStillInvoked's file anchor,
  which still pointed at SkillMetadataParser.cs after the partial-class split
  moved the method it was checking for into SkillMetadataParser.Egress.cs.
