+++ src/Content/Infrastructure/Infrastructure.AI/Skills/SkillMetadataParser.cs
@@ -2,7 +2,6 @@ using Application.AI.Common.Exceptions;
 using Application.AI.Common.Interfaces.Governance;
 using Application.AI.Common.Interfaces.Skills;
 using Application.Common.Helpers;
-using Domain.AI.Egress;
 using Domain.AI.Skills;
 using Domain.AI.Tools;
 using Domain.Common.Config.AI;
@@ -213,28 +212,6 @@ public sealed partial class SkillMetadataParser
             PluginSource = pluginSource,
             Egress = egress,
         };
-    }
-
-    /// <summary>
-    /// Validates a non-null <paramref name="egress"/> manifest against the same SSRF-narrow rules
-    /// the runtime egress policy applies (#531), throwing <see cref="SkillParsingException"/> if it
-    /// fails — a malformed manifest must never reach the policy resolver. Mirrors
-    /// <see cref="ScanOrRefuse"/>'s "refuse to construct the definition" shape for a different
-    /// class of manifest defect (schema validity, not content safety).
-    /// </summary>
-    private void ValidateEgressOrRefuse(EgressManifest? egress, string skillFilePath)
-    {
-        if (egress is null)
-            return;
-
-        var result = _egressValidator.Validate(egress);
-        if (result.IsValid)
-            return;
-
-        var reason = string.Join("; ", result.Errors.Select(e => e.ErrorMessage));
-        _logger.LogWarning(
-            "Refusing skill manifest at {Path}: invalid egress allowlist: {Reason}", skillFilePath, reason);
-        throw new SkillParsingException(skillFilePath, $"Invalid egress manifest: {reason}");
     }
 
     /// <summary>
