+++ src/Content/Tests/Infrastructure.AI.Tests/Agents/AgentMetadataRegistryTests.cs
@@ -1,6 +1,5 @@
 using Application.AI.Common.Interfaces;
 using Application.AI.Common.Interfaces.Skills;
-using Application.AI.Common.Skills;
 using Domain.Common.Config;
 using Domain.Common.Config.AI;
 using FluentAssertions;
@@ -45,7 +44,7 @@ public sealed class AgentMetadataRegistryTests
             new SkillMetadataParser(
                 NullLogger<SkillMetadataParser>.Instance, new UnsandboxedSkillFileReader(),
                 TestMcpSecurityScanner.AlwaysSafe(), TestMcpSecurityScanner.DefaultConfig(),
-                new EgressManifestValidator()),
+                TestMcpSecurityScanner.RealEgressValidator()),
             new UnsandboxedSkillFileReader(),
             ownedSkills);
     }
