+++ src/Content/Tests/Infrastructure.AI.Tests/Bundles/BundleStagingServiceTests.cs
@@ -2,7 +2,6 @@ using System.IO.Compression;
 using System.Text;
 using Application.AI.Common.Interfaces.Governance;
 using Application.AI.Common.Interfaces.Plugins;
-using Application.AI.Common.Skills;
 using Domain.AI.Governance;
 using Domain.Common.Config;
 using Domain.Common.Config.AI;
@@ -738,7 +737,7 @@ public sealed class BundleStagingServiceTests : IDisposable
                 NullLogger<SkillMetadataParser>.Instance, new UnsandboxedSkillFileReader(),
                 skillScanner ?? TestMcpSecurityScanner.AlwaysSafe(),
                 Mock.Of<IOptionsMonitor<AIConfig>>(m => m.CurrentValue == appConfig.AI),
-                new EgressManifestValidator()),
+                TestMcpSecurityScanner.RealEgressValidator()),
             new UnsandboxedSkillFileReader(),
             pluginReader ?? new PluginManifestReader(NullLogger<PluginManifestReader>.Instance),
             bundleOwnedMcpServers ?? new BundleOwnedMcpServerRegistry(),
