+++ src/Content/Tests/Infrastructure.AI.Tests/Skills/EchoTestSkillManifestTests.cs
@@ -1,4 +1,3 @@
-using Application.AI.Common.Skills;
 using Infrastructure.AI.Skills;
 using FluentAssertions;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -33,7 +32,7 @@ public sealed class EchoTestSkillManifestTests
         var parser = new SkillMetadataParser(
             NullLogger<SkillMetadataParser>.Instance, new UnsandboxedSkillFileReader(),
             TestMcpSecurityScanner.AlwaysSafe(), TestMcpSecurityScanner.DefaultConfig(),
-            new EgressManifestValidator());
+            TestMcpSecurityScanner.RealEgressValidator());
         var skillPath = RepoRoot.Combine("skills", "echo-test");
         var filePath = Path.Combine(skillPath, "SKILL.md");
