+++ src/Content/Tests/Infrastructure.AI.Tests/Skills/PluginGovernanceWiringTests.cs
@@ -2,7 +2,6 @@ using Application.AI.Common.Interfaces;
 using Application.AI.Common.Interfaces.Plugins;
 using Application.AI.Common.Interfaces.Tools;
 using Application.AI.Common.Services.Tools;
-using Application.AI.Common.Skills;
 using Domain.AI.Skills;
 using Domain.Common.Config;
 using Domain.Common.Config.AI;
@@ -83,7 +82,7 @@ public sealed class PluginGovernanceWiringTests : IDisposable
             new SkillMetadataParser(
                 NullLogger<SkillMetadataParser>.Instance, new UnsandboxedSkillFileReader(),
                 TestMcpSecurityScanner.AlwaysSafe(), TestMcpSecurityScanner.DefaultConfig(),
-                new EgressManifestValidator()),
+                TestMcpSecurityScanner.RealEgressValidator()),
             new UnsandboxedSkillFileReader(),
             pluginRegistry);
     }
