+++ src/Content/Tests/Infrastructure.AI.Tests/Skills/SkillFileReaderTests.cs
@@ -1,5 +1,4 @@
 using Application.AI.Common.Interfaces.Skills;
-using Application.AI.Common.Skills;
 using Domain.Common.Config;
 using Domain.Common.Config.AI;
 using Domain.Common.Config.AI.BundleExecution;
@@ -228,7 +227,7 @@ public sealed class SkillFileReaderTests : IDisposable
         var parser = new SkillMetadataParser(
             NullLogger<SkillMetadataParser>.Instance, reader,
             TestMcpSecurityScanner.AlwaysSafe(), TestMcpSecurityScanner.DefaultConfig(),
-            new EgressManifestValidator());
+            TestMcpSecurityScanner.RealEgressValidator());
 
         var act = () => NestedSkillScanner.Scan(
             _outsideRoot, parser, reader, NullLogger<SkillFileReaderTests>.Instance);
