+++ src/Content/Tests/Infrastructure.AI.Tests/Skills/SkillMetadataParserNestedYamlTests.cs
@@ -1,4 +1,3 @@
-using Application.AI.Common.Skills;
 using FluentAssertions;
 using Infrastructure.AI.Skills;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -19,7 +18,7 @@ public sealed class SkillMetadataParserNestedYamlTests : IDisposable
         _sut = new SkillMetadataParser(
             NullLogger<SkillMetadataParser>.Instance, new UnsandboxedSkillFileReader(),
             TestMcpSecurityScanner.AlwaysSafe(), TestMcpSecurityScanner.DefaultConfig(),
-            new EgressManifestValidator());
+            TestMcpSecurityScanner.RealEgressValidator());
         _tempDir = Path.Combine(Path.GetTempPath(), $"skill-nested-yaml-tests-{Guid.NewGuid():N}");
         Directory.CreateDirectory(_tempDir);
     }
